feat: Funnel 파일 분리 및 member 생성 api 로직 변경
